Fix the category chart when only one value exists
In case if only 1 value existed in the chart, the ordinal scale wasn't rendered properly: the calculation of the distance between the chart's columns was broken.
This fix adds an additional check for the domain: if the chart has only one value, uses the entire range is used as a distance.